Fix #154: Unable to load multiple VSTO plugins using Scintilla.NET
Under normal circumstances, a library can be loaded multiple times. This would cause the OS to optimize this by just incrementing the refcount of that library handle instead of actually loading it multiple times. This can of course be worked around by loading the exact same copy of it from a different path. Unless the library does something stupid with global mutexes/handles everything should work.

However, Scintilla devs in their infinite wisdom decided to design the library such that it registers a window class named "Scintilla" using `RegisterClassEx` with `CS_GLOBALCLASS` flag inside its `DllMain` entry point. While this allows us to actually create a "Scintilla" instance from the parent process (without the flag, only the code inside `Scintilla.dll` could create the instance), it also means we can't load 2 versions of `Scintilla.dll` because `RegisterClassEx` fails with `[0x582] ERROR_CLASS_ALREADY_EXISTS`.

This limitation could be fixed by providing an API that creates the "Scintilla" instance (handle) on behalf of the parent process, or it could provide a way to customize the class name, or it could try to find an available class name and have an API to retrieve that name instead, but these seem unlikely to happen for backwards compatibility reasons.